requirements.txt

langchain>=0.2.6
langchain-openai>=0.1.8
pydantic>=2.5
python-dotenv>=1.0.1
pytest>=8.2.0
